Implement app download functionality with improved error handling
